One of several QT files for Semantics Prototype
This is one of several QT files I created for a prototype I want to use for a paper to Semantics 2023. This takes a QT tree and uses it to display a tree of UN SDG Goals and Targets.  This needs a lot of work but it shows that this kind of hierarchy can be built in QT.