Update Daan Ontology V1.ttl
Added 3 properties from the Simple Knowledge Organization System (SKOS): skos:altLabel is for other names for an entity. skos:prefLabel is for the preferred label when there is more than one. skos:hiddenLabel is meant for strings that may be used in legacy systems (so the ontology needs to know they are names for an entity) but are never meant to be used to display the entity, i.e., they've been deprecated but still are retained for legacy systems. The order for labels is:
skos:prefLabel
rdfs:label
skos:altLabel
I.e., if there is a prefLabel it is used, otherwise the label, otherwise an altLabel. The hiddenLabel should never be used to display. The  idea is that besides being used to display entities these can also be used to find synonyms with objects in other systems.