Added File to Upload Contacts Table
This is ugly. There is a more elegant way to do it where we just read the column names and match them to the property names in the knowledge graph but I did it the simple way just to try and get it working. I list out every data property. This is mostly working. It loaded the first +10K rows in the Contacts csv file and then choked on some character that it couldn't process.