Uploaded Add Contacts
This file loads the Contacts table exported as a CSV file into the knowledge graph. This version hard codes each data property.  The readCSV.py file is a more general way to do it. It gets the names of the properties from the header of the CSV file so we don't need to write a new function for each table.