This is a more general way to load a CSV file
This version reads the first row and converts the heading into a data property. Need to do some extra error detection for when there is no existing data property (when the heading doesn't match the name of the property) but otherwise this can be used to load all future data as it doesn't hard code the data properties.