Create Daan Ontology V2
This is a new ontology with most of the original NGOs (which were from a not very good old file I had) deleted. I'm going to use this as the starting point to load the new data. I kept about 10 NGOs because they were useful for demo purposes in various papers and may still be useful. All those NGOs are instances of the class TestData as well as the other classes they would be instances of. That way if we need to, we can easily delete them all at once if needed.