Update Daan Ontology V2.ttl
Added ngoType data property to be consistent with exported data model Contacts file. Note will need to manually change "type" to "ngoType" in each Contact file. Normally, I prefer to just have the same names for the columns and properties but I worry about conflicts with rdf:type if we leave it as just "type"