Added Demo Ontology
This is a small ontology I'm using for papers. It has demo data set up to highlight specific capabilities where some data may not be available yet in the actual uploaded data. Also, since it is smaller than the actual data it is easier to manipulate, change, etc. in the future. This ontology is also useful for testing.