Added Simple SDG matching
This version just uses keyword matching. Keywords are defined for the Goals and Targets. Text of the Indicators is usually so similar to the Targets didn't think it was worth including them in this version (plus it's some non-trivial work to add the keywords in a way that seems likely to avoid both Type I (false positive) and Type II (false negative) errors.